Change Function of timestamp in file
Initial timestamp is now written to the top of the file (in nanoseconds).
The timestamp entry now corresponds to the "sample number" (entries will start with 1 and increment by 1 for each additional entry).

This change was to counteract the loss of precision in MATLAB caused by the huge timestamp numbers.